move SVGs into component dir, bc why not?
