+++ src/components/PageSections/Footer/index.js
@@ -4,9 +4,9 @@ import { FormattedMessage, defineMessages, injectIntl } from 'react-intl';
 import ExternalLink from 'components/ExternalLink';
 import ThreeOneOne from 'components/PageSections/ThreeOneOne';
 
-import TwitterSVG from 'js/svg/Twitter';
-import FacebookSVG from 'js/svg/Facebook';
-import GithubSVG from 'js/svg/Github';
+import TwitterSVG from 'components/SVGs/Twitter';
+import FacebookSVG from 'components/SVGs/Facebook';
+import GithubSVG from 'components/SVGs/Github';
 import citySealImg from 'images/coa_seal_white.png';
 
 const i18nMessages = defineMessages({
