+++ src/components/SVGs/CaretUp.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import SVG from 'js/svg/Svg';
+import SVG from 'components/SVGs/Svg';
 
 const CaretUp = ({ title="Arrow Up", ...rest }) => (
   <SVG
