+++ src/components/SVGs/Twitter.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import SVG from 'js/svg/Svg';
+import SVG from 'components/SVGs/Svg';
 
 const TwitterSVG = ({ title="Twitter", ...rest }) => (
   <SVG
