+++ src/components/Tiles/Tile.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import I18nLink from 'components/I18nLinks/I18nLink'
-import ChevronCircleRightSVG from 'js/svg/ChevronCircleRight';
+import ChevronCircleRightSVG from 'components/SVGs/ChevronCircleRight';
 
 const Tile = ({ url, text, tag }) => (
   <I18nLink className="coa-Tile" to={url}>
