+++ src/components/Tiles/TileGroup.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import Tile from 'components/Tiles/Tile';
 import SectionHeader from 'components/SectionHeader';
 import I18nLink from 'components/I18nLinks/I18nLink';
-import ArrowRight from 'js/svg/ArrowRight';
+import ArrowRight from 'components/SVGs/ArrowRight';
 
 
 const TileGroup = ({ tiles, tag, title, titlePath, description, hasBorder }) => (
